Move softmaxgrad dnn test to test_dnn.py file
